backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
requests
ortools
networkx
python-dotenv
websockets
jinja2
